+++ client/src/index.jsx
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom/client'
 import App from './app/app'
 import Login from './components/pages/login/login'
 import Dashboard from  './components/pages/dashboard/dashboard';
-import SignUp from './components/sign_up/sign-up'
-import ForgotPassword from './components/forgot_password/forgot-pasword'
+import SignUp from './components/pages/sign_up/sign-up'
+import ForgotPassword from './components/pages/forgot_password/forgot-pasword'
 import {
   createBrowserRouter,
   createRoutesFromElements,
